\docType{data}
\name{AngKrug91}
\alias{AngKrug91}
\title{Data used in Angrist and Krueger 1991}
\format{A data frame containing 329509 observations and 5 variables.}
\source{
TODO
}
\usage{
AngKrug91
}
\description{
TODO
}
\details{
TODO
}
\references{
Krueger, A. (1999) "Experimental Estimates Of Education
Production Functions," \emph{The Quarterly Journal of
Economics}, Vol. 114(2), pages 497-532, May.

Angrist, A. and Krueger, A. (1991) Does Compulsory
Schooling Attendance Affect Schooling and Earnings?,
\emph{Quarterly Journal of Economics}, 106, 976-1014.
}

